Extract /categories/... to separate module
