+++ app/api/routes/__init__.py
@@ -8,7 +8,7 @@ from algoliasearch.exceptions import AlgoliaException, AlgoliaUnreachableHostExc
 from app import Config, db, index
 from app.api import bp
 from app.api.auth import authenticate
-from app.api.routes import api_key
+from app.api.routes import api_key, categories
 from app.api.routes.logger import logger
 from app.api.routes.metrics import latency_summary, failures_counter
 from app.api.validations import (
@@ -112,20 +112,6 @@ def languages():
 @bp.route('/languages/<int:id>', methods=['GET'], endpoint='get_language')
 def language(id):
     return get_language(id)
-
-
-@latency_summary.time()
-@failures_counter.count_exceptions()
-@bp.route('/categories', methods=['GET'])
-def categories():
-    return get_categories()
-
-
-@latency_summary.time()
-@failures_counter.count_exceptions()
-@bp.route('/categories/<int:id>', methods=['GET'], endpoint='get_category')
-def category(id):
-    return get_category(id)
 
 
 # Helpers
@@ -310,36 +296,6 @@ def get_language(id):
 
     if language:
         return utils.standardize_response(payload=dict(data=(language.serialize)))
-
-    return redirect('/404')
-
-
-def get_categories():
-    try:
-        category_paginator = utils.Paginator(Config.CATEGORY_PAGINATOR, request)
-        query = Category.query
-
-        paginated_categories = category_paginator.paginated_data(query)
-        if not paginated_categories:
-            return redirect('/404')
-        category_list = [
-            category.serialize for category in paginated_categories.items
-        ]
-        pagination_details = category_paginator.pagination_details(paginated_categories)
-    except Exception as e:
-        logger.exception(e)
-        return utils.standardize_response(status_code=500)
-
-    return utils.standardize_response(payload=dict(
-        data=category_list,
-        **pagination_details))
-
-
-def get_category(id):
-    category = Category.query.get(id)
-
-    if category:
-        return utils.standardize_response(payload=dict(data=(category.serialize)))
 
     return redirect('/404')
